added 2 new functions
getArray to seperate all of the words in the string into their own array item. adds the array item to the matchWords array if one of its letters matches the target letter. not working. i think the problem is on line 36. I am using indexOf on an array instead of a string within in an array-- maybe.